requirements.txt
beautifulsoup4==4.12.2
nltk==3.6.5
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.12.0
altair==4.1.0
plotly==5.7.0
